Do not use EV_ONESHOT when monitoring the dhcpd leases file.
If the dhcpd daemon rewrite the leases file and we reopen it too quickly we
can miss events because kqueue may not have the chance to flush its event
queue (i.e. EV_ONESHOT is still set and will prevent us from receive further
events for leases file).

This fixes the issue #4931