patch_remove_shellcmd() removal count logged
Presumably the number of occurances removed should always be 0 or 1, but if it's ever more then odd behaviour could be hard to track down, especially as all we match on is the fragment "apply_patches.php which could match other commands unwittingly in future. So this logs how many removals were done (even though should be redundant) because it won't hurt to be precise with an actual count.