MFH: r405175
[Backporting this change is necessary after r409158 introduced a requirement
on this port in the 2016Q1 branch]

- Add p5-Schedule-Cron-Events 1.94
- While I'm here, add NO_ARCH

Given a line from a crontab, tells you the time at which cron will next run the
line, or when the last event occurred, relative to any date you choose. The
object keeps that reference date internally, and updates it when you call
nextEvent() or previousEvent() - such that successive calls will give you a
sequence of events going forward, or backwards, in time.

Use setCounterToNow() to reset this reference time to the current date on your
system, or use setCounterToDate() to set the reference to any arbitrary time, or
resetCounter() to take the object back to the date you constructed it with.

This module uses Set::Crontab to understand the date specification, so we should
be able to handle all forms of cron entries.

WWW: http://search.cpan.org/dist/Schedule-Cron-Events/

PR:		205247
Submitted by:	Andrey Kuzmin <akuz84@gmail.com>

PR:		205867
Approved by:	portmgr (antoine)