MFH: r412524
During the exp-run in bug 208158, it was found that comms/telldus-core
gives errors with libc++ 3.8.0 [1]:

/wrkdirs/usr/ports/comms/telldus-core/work/telldus-core-2.1.2/service/ProtocolNexa.cpp:59:8: error: call to member function 'insert' is ambiguous
                        str.insert(0, 1, 2);  // Repeat two times
                        ~~~~^~~~~~

This is because the third argument to this std::string method is
supposed to be a char, not an integer.  Fix it by using '\x2' instead.

[1] http://package18.nyi.freebsd.org/data/headamd64PR208158-default/2016-03-22_18h30m05s/logs/errors/telldus-core-2.1.2_1.log

Approved by:	portmgr (junovitch)
Approved by:	johan@stromnet.se (maintainer)
PR:		208523