Automatically update apinger to dpinger
in the Service Watchdog watch list.
Forum: https://forum.pfsense.org/index.php?topic=109770.msg611158#msg611158

Doing it here means that people will get the automatic change as soon as they upgrade to 2.3 and the upgrade installs the Service Watchdog code from here. The other place it could be done is in upgrade_config.inc - but then that will not be distributed to them until pfSense 2.3.1 happens - rather too late.