Status / Monitoring - Last Sample Zero
Monitoring graphs last sample being zero

If there is no RRD data sample for the current minute then the graph and last and minimum data value will be zero in certain circumstances.  Such as resolution 1 minute and graph loaded prior to the current minute RRD sample being made.  Similar for the first minute of other resolutions.

This commit corrects this situation by never requesting RRD data for the uncompleted current minute to ensure that there is a data point for the last/end time.