Update for siproxd-0.8.2
Updated the config page to support most available settings in
siproxd 0.8.2, including all the new plugins.
Also updated the show "Registered Phones" page to display the
registered UAs properly as the storage file format did change
a bit (0.8.1 -> 0.8.2).

This package MUST go along with siproxd-0.8.2 (as is present in
the FreeBSD ports tree).