Small cosmetic patch to System Patches list view
When viewing installed patchlist on 2.3+ the commit URLs are quite long and cause the main table to extend beyond the screen boundary. This looks weird. Seeing the commit URL on the list doesn't add too much as far as usability and it can be seen anyway if needed on the edit/detail page. So this patch removes it and I think the output looks much cleaner.