Fix description on the Management option for LADVD.
The "management" interface does not work on VLAN or LAGG interfaces, but requires a physical interface with an IP address. Automatically using the parent likely would hit an interface that has no address, so there is no good automated way to solve this currently.

When an improper interface is chosen, LADVD will (harmlessly) spam the log with "ladvd: could not detect the specified management interface"