add option to log payload data to eve.json alerts
add option to log payload data to eve.json alert logging. This could probably use a check to make sure alert logging is on if alert payload logging is selected. as it is now it does no alert logging if alert payload is selected and alert is not selected
(cherry picked from commit 4b460b855c0551cd5497266701bb60ea8991e608)