Created no-release '-n' option.
no-release option added to prevent a release signal being sent to the ISP. Some ISP's will release the given allocation and issue a new allocation on next connect, regardless of the DUID. As this is a default state for dhcp6c whenever it terminates it has to be prevented. The -n option prevents the signal from ever being sent.

Some LOG_DEBUG info has been changed to LOG_INFO and some extra logging added also.