Added -n no release option & extra info logging
No release option '-n' added to prevent dhcp6c from sending release signal to ISP's. Some ISP's will release the allocation on receipt of the signal and issue a new PD and/or address on next request. As this is the default when dhcp6c exits the only way to stop it is to prevent it from being sent. Also added some extra information to the default logging so the process can be seen without having to select debug logging.