dhcp6c no release option and extra logging
Added a no release '-n' option to dhcp6c. This is to prevent some ISP's will release the allocated PD or address on receipt of the release signal and issue a new address on next request. As dhcp6c by default sends the release on exit the only way to stop it is by stopping all release signals being sent. the -n option does that. Also added is some extra info logging which is minimul and useful without resorting to the debug log.