Added no release option & additional light logging
Add -n no release option to prevent dhcpc6 from sending release signal to ISP. Some ISP's will release the allocated address or PD on receipt of the signal and issue new ones on the next request. As this is the default behavior of dhcp6c the only option is to stop it from sending release signals completely, this option handles that used in conjunction with an option in WAN DHCP6 configuration. Also added is some light extra INFO logging that helps to show what dhcp6c is doing.