Added new flag '-x' prevents script execution
In pfsense, dhcp6c will launch the wan script on receipt of a reply from the dhcp6 server. When in dhcp6withoutRA mode, that same script will also get called on receipt of the RA signal by rtsold. This is not an optimal situation with the same script running twice at the same time. The additional flag prevents dhcp6c from running the script only when it receives a REPLY signal from the server. As the REPLY signal is only sent when replying to a REQUEST, and as the REQUEST is only sent at the start of a session then any other trigger to run the script will execute it as normal.