MFH: r431537
Add "dual dhclient" script, which can be specified as dhclient_program in
rc.conf in order to launch the base system dhclient for IPv4 and the ISC
dhcp client for IPv6.

I'm not sure if this will be useful to anyone else, but adding this trivial
port will simplify the configuration of IPv6 on Amazon EC2 instances.  I
hope this port becomes redundant in the future due to the base system
gaining support for IPv6 DHCP.

Approved by:	portmgr (mat)