New feature for the SID Mgmt: rule content modification by category
This commit adds functionality to modify all signatures within a category.

suricata_modify_sid_content was restructured to walk always over the sid list. Now it is actually a list is of sids,category and the wildcard '*'. If an item is a specific SID it gets modified otherwise the whole rule_map is walk thru to find all rules that need be changed.