Redmine #7548 Add absolute offset to NTP stats
Plot and list the absolute value of the NTP time offset, as well as the existing plus/minus value.
See Redmine https://redmine.pfsense.org/issues/7548
Forum https://forum.pfsense.org/index.php?topic=130468.0
for details.

In order to do add this, it was easier to also refactor the whole code to get rid of the various counters and huge block of duplicated code for left and right. Those refactorings are in the commits before this one.