Disable download site carrying botched tarball.
The smaller tarball at build.openvpn.net is a pre-release version that
somehow escaped to the public and would not carry the fix for
CVE-2017-7478.

I had grabbed the bigger tarball from swupdate.openvpn.net that does
carry the fix, so we're safe.

The bigger tarball does carry garbage files (pre-compiled .so files, which
we do not install or use in the build, and *~ backup files) which do not
end up in our build.

The issue has been reported upstream.

If someone gets download failures due to the size differences, "make
distclean" should fix things.

This commit invokes the "Fixes that do not result in a change in
contents of the resulting package." blanket approval per
<https://wiki.freebsd.org/ports-secteam>, checked today.

Reported by:	garga@
Approved by:	ports-secteam@ (blanket)