pfSense-pkg-suricata: Delete rotated `eve.json` log files
The automatic log removal under `suricata_check_dir_size_limit` does not clean-up the ève.json.*` log files properly. This can lead to a situation where the other rotated logs are cleared but the overall directory size can still grow beyond the hard limit. 

For example, if `eve.json` log retention is set to more than a few days and with `flow` logging enabled: there you can end up with dozens of `32MB` files that are not cleared in time.  

With this change, that shouldn't happen anymore.