MFH: r458014
devel/libfmt: Bump PORTREVISION

Attempt to fix build of kodi-devel on 2018Q1.

r453704 moved libfmt from building a static library to a shared
library but PORTREVISION was not bumped.

This now leads to build failures on 2018Q1 where libfmt.so cannot be
found because the libfmt package was not rebuilt and only installs
libfmt.a.

http://beefy1.nyi.freebsd.org/data/111i386-quarterly/457961/logs/kodi-devel-18.0.g20171213_1.log
http://beefy3.nyi.freebsd.org/data/111amd64-quarterly/457961/logs/kodi-devel-18.0.g20171213_1.log

Pointy hat:	yuri

Approved by:	ports-secteam (blanket)