MFH: r457853
devel/awscli: use USE_PTYHON=noflavors, remove PKGNAMEPREFIX

The introduction of flavors caused the awscli package to be renamed to
py27-awscli. Since this package is preinstalled on all AWS EC2 instances
running a FreeBSD image, apply POLA and rename the package back to
awscli. The assumption is that the rename was unintentional.

Approved by:	mat (co-mentor)
Differential Revision:	https://reviews.freebsd.org/D13555

Approved by:	portmgr