AutoMagic: Create remote lines for every port within an alias
When you have a Multi-WAN OpenVPN server or multiple ports NATing to a single OpenVPN server, many people recommend using an alias for the ports. Unfortunately the old implementation only creates a remote line for the first port of the alias and discards the rest. Generating multiple remote lines in this case is the expected behaviour.