Properly check domain when registering DHCP static mappings
Use of the PHP function "array_diff_assoc" results in any zone which is
a subdomain of a given static DHCP mapping domain being populated with
said mapping. In other words, a static DHCP mapping with a domain of
"domain.com" would appear in any zone that is a subdomain of it, such
as "sub1.domain.com" or "sub2.sub1.domain.com".

This fix changes the logic to perform a direct domain comparison so
that only the zone matching the exact domain specified in a given
static DHCP mapping will be populated with such a mapping entry.