Add a patch to fix OSMAJOR poudriere ARM jails
When build host runs FreeBSD 12 amd64 and jail is FreeBSD 11 armv6 or
aarch64 using native-xtools __FreeBSD__ macro, that comes from the
compiler, is set to 12. It makes pkg believes there is a major OS
update and print the warning about it on every single call.

My first attempt was to use (__FreeBSD_version / 100000), because
__FreeBSD_version comes from sys/param.h and would use the header from
jail. It didn't work because of INDEXFILE macro declaration that uses
OSMAJOR to concatenate to "INDEX-" prefix and preprocessor wouldn't do
the math before replace that.

After checking that uname from jail is not being replaced by
native-xtools, and I believe it will never be, because it would lead
scripts to consider they are on a 12 system, I chose this path of using
a autoconf knob OS_MAJOR collecting data from uname output.