Remove standard RADIUS attributes
Radius Attributes are currently not used in a common/standard way by pfSense Authentication components (currently OpenVPN, IPsec xAuth and CaptivePortal) when performing an Authentication.
In an attempt to clarify RADIUS PHP code, i suggested in pfsense/pfsense#3640 to standardize the way attributes are used, but outside of the RADIUS class, in auth.inc. This would make normalization easier, and also make the code more clear (i am not sure that using functions like getNasIp() to  retrieve arguments is a best practice. Shouldn't arguments be used instead)?