MFH: r489642 r489971 r489992 r490212
Deprecate net/GeoIP, and expire it at the end of the month

Note that only the legacy format is deprecated. The GeoIP2 databases
are alive and well.

I completely missed that GeoIP has been deprecated for the last year,
and it expired upstream last week. Maxmind no longer provides the
legacy database, rendering this port useless.

Users must switch to GeoIP2 immediately. OPTIONS changes and
instructions will be coming in the next few days.

See https://support.maxmind.com/geolite-legacy-discontinuation-notice/

Although it won't be purged from the quarterly branch, it probably
makes sense to MFH this deprecation notice.

PR:             234715
Reported by:    rigoletto

Replace the geoipupdate.sh script (which just produced 404 errors) with
a message explaining why, and what to do about it.

Report failure to the caller in geoipupdate.sh

The script was returning an error due to the service not being offered anymore.
The update attempt was replaced with an information message, but it's better to
still return 1 as it did before to allow users detecting they have a problem.

Reported by:	aramw (maintainer)

Extend the GeoIP expiration to a month from now to give
users a bit more time to find a new solution.

Approved by:	portmgr (with hat)