MFH: r497494
Fix gexiv dependencies in the port.

The port was erroneously depending on graphics/py-exiv2 while in fact it only
needs exiv2 itself (which was always present anyway because py-exiv2 depends on
it). This port builds a Python module that links against exiv2, so clarify the
situation and stop depending on py-exiv2, which is BROKEN and DEPRECATED.

PR:		236806
Approved by:	J.R. Oldroyd <fbsd@opal.com> (maintainer)

Approved by:	ports-secteam (blanket approval)