pfSense-upgrade: Fix proxy usagee on FreeBSD > 11
A hack was added in the past to prevent proxy user/pass variables to be
set on FreeBSD 10 due to a bug on libfetch.  This hack was adding these
variables using a check strict to FreeBSD version 11.

Since we don't support upgrade direct from 2.3 (FreeBSD 10) to 2.5
(FreeBSD 12), remove this hack and let the proxy to be used.

It should fix ticket #9478