MFH: r518116
Use clang 7 as compiler on 11.x and 12.x to fix build.

clang 8 and later error for duplicate definitions of enum values.  mdb
depends on having identical but duplicate definitions.  This was fixed
differently in 13 with a change to the kernel, but that change is
disruptive and cannot be easily backported.

Reported by:	pkg-fallout, Attila Nagy
Approved by:	bdrewery
Differential Revision:	https://reviews.freebsd.org/D22486

Approved by:	ports-secteam (joneum)