pfSense-upgrade: Upgrade itself and re-run
pfSense-upgrade was being registered as a dependency of pfSense-repo
pkg.  Because of that, when pfSense-repo package was upgraded to latest
version, it was upgraded together and the code used to detect a new
version of pfSense-upgrade was not being triggered.

The dependency was wrong and was removed, and added to the proper place,
security/pfSense.  But it cannot be changed on older versions.

Add a workaround to detect, before upgrade pfSense-repo, if a newer
version of pfSense-upgrade is available, and when it happens, make it to
exit expected code, 99, so wrapper can restart the process.