Fix install-missing-packages when port is missing multiple dependencies
The current implementation of install-missing-packages does not handle
correctly the situation when a port is missing multiple dependencies. pkg(8)
would only get the first missing origin as an argument. All the other origins
would be listed on new lines and would not be passed to pkg(8).

PR:		244215
Discussed with:	mat, tobik
Approved by:	portmgr (implicit, fix r525388)
Differential Revision:	https://reviews.freebsd.org/D23795