New port: multimedia/makemkv a CLI video converter from DVD
This port is CLI only and leaves out the MakeMKV GUI, because
makemkvcon (closed source Linux binary) won't run in guiserver
mode.

Also, as described in the pkg-message, it requires a custom
FreeBSD kernel with the Linux SCSI generic drivers and has to
be build by the final user to comply with the license terms.

Work is ongoing with the author to have a FreeBSD native version
of this port in the future.

PR:		238004
Submitted by:	Felix Palmen <felix@palmen-it.de>