devel/softether-devel: Update to v4.34.9745-beta
Fix security issue:
Fix the security of JSON-API. If the administrator password of the
VirtualHub is empty, JSON-API (which was added in 4.30 Build 9696 Beta)
will not be able to access to the virtual hub with a empty password since
this release. Because there are relatively many cases in which administrator
password is empty for a virtual hub, being able to manage a virtual hub
without a password using JSON-API was a security problem. In this release,
this behavior has been changed so that JSON-API cannot be accessed in the
Virtual Hub management mode until it is configured with non-empty password.

Relnotes:	https://www.softether.org/5-download/history
Sponsored by:	HAW International