MFH: r531528 Approved by:	portmgr (with hat)
vim: Fix environment contamination from libcanberra

Vim will link against libcanberra if it's present, leading to an
undeclared link and breakage risk if canberra is removed. It only really
makes sense for it to do this during a gnome build, so explicitly
enable it there and disable it elsewhere.

PR:		245460
Reported by:	Andy Mender