* Updated to 7.1.3
This minor version update no longer supports python 2.7.
The way mysql support is detected upstream has changed. Instead of patching
the configure.ac script, I used a couple hacks for this to be detected for
custom configurations. Didn't detect any issues with this method but I'm
open to change this in the future.

Added a SeaHub upgrade message to smooth out upgrading. For upgrading from 7.0.x,
if one does not clear the cache, http 500 will be returned.

The seafile_datadir no longer points to the seahub.ini, this has been changed
upstream and may cause upgrading issues if not using the default. This can
easily be changed by setting it in rc.conf.

Changes: https://download.seafile.com/published/seafile-manual/changelog/changelog-for-seafile-professional-server.md