Remove the old emulators/qemu-user-static from the tree
This will be momentarily replaced by a repocopy of qemu-user-static-devel
back over to qemu-user-static. With the update of qemu-user-static to 3.1,
emulators/qemu-user-static will no longer be a slave port and
emulators/qemu-sbruno will be deprecated in a couple months.

Some history is being lost in this transition, but almost all of the
interesting parts of qemu-user-static were in the master port qemu-sbruno
anyways, which qemu-user-static-devel is repoderived from.

PR:		243906
Approved by:	koobs (mentor)