Mark the port as requiring postgresql-11 or higher.  Attempting to build the port with postgres-10 results in:
```
configure: error: Citus is not compatible with the detected PostgreSQL version 10.
```

No PORTREVISION bump, as this will have no effect on the default
packages (which link against postgresql-11) or on anyone who has
successfully built pg_citus.  All that happens is that we detect the
incompatible postgresql version earlier in the build process and error
out sooner.

PR:		245958
Reported by:	rainer@ultra-secure.de